nuevas pantallas y room añadido

+++ frontend-android/PrototipadoComunidades/proyectobase2425/app/src/main/java/com/pmdm/proyectobase2425/ui/features/comunidadesDentro/ComunidadesDentroViewModel.kt
@@ -14,7 +14,7 @@ class ComunidadesDentroViewModel @Inject constructor() : ViewModel() {
     var state by mutableStateOf(ComunidadesDentroUiState())
         private set
 
-    fun cargarComunidad(comunidadId: Int) {
+    fun cargarComunidad(comunidadId: Long) {
         // TODO: cargar comunidad por ID desde repositorio
         state = state.copy(isLoading = false)
     }
