+++ frontend-android/PrototipadoComunidades/proyectobase2425/app/src/main/java/com/pmdm/proyectobase2425/ui/navigation/AppNavigation/route/EventoRoute.kt
@@ -12,7 +12,7 @@ import androidx.navigation.toRoute
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class EventoRoute(val eventoId: Int)
+data class EventoRoute(val eventoId: Long)
 
 fun NavGraphBuilder.eventoDestination(
     navController: NavHostController
